the-continental-bot/requirements.txt
discord.py>=2.3.2,<3.0.0
aiohttp>=3.8.6,<4.0.0
python-dotenv>=1.0.0,<2.0.0
cloudinary>=1.36.0,<2.0.0